# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Infinite Local API
  version: 0.15.5
  description: Loopback-only API used by the Agent Infinite renderer.
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port: { default: '0' }
security:
  - bearerAuth: []
paths:
...
  /api/models/scan:
    post:
      operationId: scanModels
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                provider: { type: string, enum: [claude, codex, pi, opencode] }
      responses:
        '200':
          description: Refreshed model catalog
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModelInventory' }
        '422': { $ref: '#/components/responses/ErrorResponse' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
    hookSessionAuth: { type: apiKey, in: header, name: X-Agent-Infinite-Hook-Token }
...
  responses:
    ErrorResponse:
      description: Request failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [code, message, details]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    ModelInventory:
      type: object
      required: [providers, scannedAt]
      properties:
        scannedAt: { type: string, format: date-time }
        providers:
          type: array
          items: { $ref: '#/components/schemas/ProviderModelCatalog' }
    ProviderModelCatalog:
      type: object
      required: [provider, status, scannedAt, models]
      properties:
        provider: { type: string, enum: [claude, codex, pi, opencode] }
        cliVersion: { type: string }
        defaultModel: { type: string }
        defaultSource: { type: string }
        status: { type: string, enum: [ok, stale, scan_error] }
        error: { type: string }
        scannedAt: { type: string, format: date-time }
        models:
          type: array
          items: { $ref: '#/components/schemas/ProviderModel' }
    ProviderModel:
      type: object
      required: [id, source, status]
      properties:
        id: { type: string }
        displayName: { type: string }
        source: { type: string }
        status: { type: string, enum: [available, missing, unverified] }
        isDefault: { type: boolean }
...